+++ src/test/java/com/moandjiezana/toml/BareKeysTest.java
@@ -2,14 +2,9 @@ package com.moandjiezana.toml;
 
 import static org.junit.Assert.assertEquals;
 
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 
 public class BareKeysTest {
-  
-  @Rule
-  public final ExpectedException exception = ExpectedException.none();
   
   @Test
   public void should_ignore_spaces_around_key_segments() throws Exception {
@@ -18,32 +13,23 @@ public class BareKeysTest {
     assertEquals("a", toml.getString("a.b.c.key"));
   }
 
-  @Test
+  @Test(expected = IllegalStateException.class)
   public void should_fail_when_characters_outside_accept_range_are_used_in_table_name() throws Exception {
-    exception.expect(IllegalStateException.class);
-    exception.expectMessage("Invalid table definition: [~]");
-    
     new Toml().parse("[~]");
   }
 
-  @Test
+  @Test(expected = IllegalStateException.class)
   public void should_fail_when_characters_outside_accept_range_are_used_in_key_name() throws Exception {
-    exception.expect(IllegalStateException.class);
-    
     new Toml().parse("~ = 1");
   }
 
-  @Test
+  @Test(expected = IllegalStateException.class)
   public void should_fail_on_sharp_sign_in_table_name() throws Exception {
-    exception.expect(IllegalStateException.class);
-
     new Toml().parse("[group#]\nkey=1");
   }
   
-  @Test
+  @Test(expected = IllegalStateException.class)
   public void should_fail_on_spaces_in_table_name() throws Exception {
-    exception.expect(IllegalStateException.class);
-
     new Toml().parse("[valid  key]");
   }